feat: add round UI helper
