+++ tests/helpers/classicBattle/autoSelect.test.js
@@ -87,7 +87,8 @@ describe("classicBattle auto select", () => {
     const battleMod = await import("../../../src/helpers/classicBattle.js");
     const store = battleMod.createBattleStore();
     battleMod._resetForTest(store);
-    await battleMod.startRound(store);
+    const { roundNumber } = await battleMod.startRound(store);
+    battleMod.applyRoundUI(store, roundNumber);
     timerSpy.advanceTimersByTime(31000);
     await vi.runOnlyPendingTimersAsync();
     const events = dispatchSpy.mock.calls.map((c) => c[0]);
@@ -118,7 +119,8 @@ describe("classicBattle auto select", () => {
     const battleMod = await import("../../../src/helpers/classicBattle.js");
     const store = battleMod.createBattleStore();
     battleMod._resetForTest(store);
-    await battleMod.startRound(store);
+    const { roundNumber: roundNumber2 } = await battleMod.startRound(store);
+    battleMod.applyRoundUI(store, roundNumber2);
     timerSpy.advanceTimersByTime(31000);
     await vi.runOnlyPendingTimersAsync();
     const events = dispatchSpy.mock.calls.map((c) => c[0]);
