+++ tests/helpers/classicBattle/interrupt.test.js
@@ -93,7 +93,8 @@ describe("classicBattle interrupts", () => {
     const battleMod = await import("../../../src/helpers/classicBattle.js");
     const store = battleMod.createBattleStore();
     battleMod._resetForTest(store);
-    await battleMod.startRound(store);
+    const { roundNumber } = await battleMod.startRound(store);
+    battleMod.applyRoundUI(store, roundNumber);
     timerSpy.advanceTimersByTime(31000);
     await vi.runOnlyPendingTimersAsync();
     const events = dispatchSpy.mock.calls.map((c) => c[0]);
