+++ tests/helpers/classicBattle/scheduleNextRound.test.js
@@ -112,12 +112,14 @@ describe("classicBattle scheduleNextRound", () => {
     const store = battleMod.createBattleStore();
     battleMod._resetForTest(store);
     const startRoundWrapper = vi.fn(async () => {
-      await battleMod.startRound(store);
+      const { roundNumber } = await battleMod.startRound(store);
+      battleMod.applyRoundUI(store, roundNumber);
     });
     await orchestrator.initClassicBattleOrchestrator(store, startRoundWrapper);
     const machine = orchestrator.getBattleStateMachine();
 
-    await battleMod.startRound(store);
+    const { roundNumber } = await battleMod.startRound(store);
+    battleMod.applyRoundUI(store, roundNumber);
 
     machine.current = "roundOver";
     await orchestrator.dispatchBattleEvent("continue");
